requirements.txt
python-dotenv
openai
playsound3